requirements.txt
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0
accelerate>=0.24.0
qdrant-client>=1.6.0
roqoqo-python>=0.12.0
numpy>=1.24.0
scipy>=1.11.0
qutip>=0.5.0
tabulate>=0.9.0
pyo3>=0.20.0
petgraph>=0.6.0
uuid>=1.0.0
chrono>=0.4.0
serde>=1.0.0
serde_json>=1.0.0
rust_brain>=0.1.0
